pyproject.toml
[project]
name = "staffsergeant"
dynamic = ["version"]
requires-python = ">=3.12"

dependencies = [
    "Jinja2",
    "linkify-it-py",
    "markdown-it-py",
    "MarkupSafe",
    "mdit-py-plugins",
    "mdurl",
    "PyYAML",
    "uc-micro-py",
    "typer",
]

[project.optional-dependencies]
dev = ["ruff", "pytest"]

[project.scripts]
ssg = "staffsergeant.ssg:main"

[build-system]
requires = ["setuptools>=80", "setuptools_scm[simple]>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
[tool.setuptools.packages.find]
include = ["staffsergeant*"]

[tool.ruff]
exclude = ["build"]


[tool.ruff.lint]
select = ["E", "F"]
